% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{avar}
\alias{avar}
\title{Computes average value at risk}
\usage{
avar(value, reference_dst, alpha)
}
\arguments{
\item{value}{Random variable (as a vector over realizations)}

\item{reference_dst}{Reference distribution of the same size as value}

\item{alpha}{Confidence value. 0 is worst case, 1 is average}
}
\value{
A list with dst as the distorted distribution,
         and value as the avar value
}
\description{
Computes average value at risk
}
