% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_qvalues}
\alias{compute_qvalues}
\title{Computes the function for the MDP for the given value function and discount factor}
\usage{
compute_qvalues(mdp, discount, valuefunction)
}
\arguments{
\item{mdp}{A dataframe representation of the MDP. Each row
represents a single transition from one state to another
after taking an action a. The columns are:
idstatefrom, idaction, idstateto, probability, reward}

\item{discount}{Discount factor in [0,1]}

\item{valuefunction}{A dataframe representation of the value function. Each row
represents a state. The columns must be idstate, value}
}
\value{
Dataframe with idstate, idaction, qvalue columns
}
\description{
Computes the function for the MDP for the given value function and discount factor
}
