% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gurobi_set_param}
\alias{gurobi_set_param}
\title{Sets a gurobi parameter. Even numeric values may be provided as strings}
\usage{
gurobi_set_param(optimizer, param, value)
}
\arguments{
\item{optimizer}{Which internal gurobi environment should be used. Should be one of
c('nature', 'lp', 'nonconvex', 'other'). Any other string leads to an error.}

\item{param}{The name (string) of the parameter to set}

\item{value}{String value of the parameter (see examples)}
}
\description{
See https://www.gurobi.com/wp-content/plugins/hd_documentations/documentation/9.0/refman.pdf
for examples
}
\details{
For example, to enable logging call:
gurobi_set_param("OutputFlag", "1")
}
\examples{
gurobi_set_param("TimeLimit", "100.0")

}
