% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_mdp_transition}
\alias{matrix_mdp_transition}
\title{Constructs transition probability matrix the MDP}
\usage{
matrix_mdp_transition(mdp, policy)
}
\arguments{
\item{mdp}{A dataframe representation of the MDP. Each row
represents a single transition from one state to another
after taking an action a. The columns are:
idstatefrom, idaction, idstateto, probability, reward}

\item{policy}{The policy used to construct the transition probabilities and rewards.
It can be a deterministic policy, in which case it should be a dataframe
with columns idstate and idaction. Both indices are 0-based.
It can also be a randomized policy, in which case it should be a dataframe with
columns idstate, idaction, probability.}
}
\value{
A list with P and r, the transition matrix and the reward vector
}
\description{
The method constructs the transition probability matrix (stochastic matrix)  P_pi
and rewards r_pi for the MDP
}
