% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdp_clean}
\alias{mdp_clean}
\title{Cleans the MDP dataframe}
\usage{
mdp_clean(mdp)
}
\arguments{
\item{mdp}{Dataframe representation of the MDP, with columns
idstatefrom, idaction, idstateto, probability, reward}
}
\description{
Makes cosmetic changes to the MDP. It mostly aggregates transition probabilities
to the same state. When there are multiple rows that represent the same transition,
then it sums the probabilities and computes a weighted average of the rewards.
}
\details{
The method achieves this by parsing and de-parsing the MDP definition.
}
