% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdp_from_samples}
\alias{mdp_from_samples}
\title{Builds an MDP from samples}
\usage{
mdp_from_samples(samples_frame)
}
\arguments{
\item{samples_frame}{Dataframe with columns idstatefrom, idaction, idstateto, reward}
}
\description{
Builds an MDP from samples
}
