% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdp_population}
\alias{mdp_population}
\title{Creates an MDP for a pest population control problem}
\usage{
mdp_population(
  capacity,
  initial,
  growth_rates_exp,
  growth_rates_std,
  rewards,
  external_mean,
  external_std,
  s_growth_model
)
}
\arguments{
\item{capacity}{Maximum possible carrying capacity (or maximum population)}

\item{initial}{Initial population (for simulation purposes)}

\item{growth_rates_exp}{A matrix of expected growth rates (also known as r_0) for
each action (row) and pest population (column).}

\item{growth_rates_std}{A matrix of standard deviations for the growth rates for
each action (row) and pest population (column). The actuall growth
is sampled from the Normal distribution.}

\item{rewards}{A matrix for the reward for taking each action in the particular population level
for each action (row) and population (column)}

\item{external_mean}{Mean external pest immigration in each period.}

\item{external_std}{Standard deviation for the external pest immigration. Normally distributed.}

\item{s_growth_model}{One of `logistic` or `exponential` representing the interaction between
the pest population and the carrying capacity (see wikipedia)}
}
\description{
The state is the pest population and action is a possible control intervention (pesticide).
Several possible population growth models are available, and the effectiveness of
the control actions is configurable.
}
