% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pack_actions}
\alias{pack_actions}
\title{Packs MDP actions to be consequtive.}
\usage{
pack_actions(mdp)
}
\arguments{
\item{mdp}{Dataframe representation of the MDP, with columns
idstatefrom, idaction, idstateto, probability, reward}
}
\description{
If there is a state with actions where idaction = 0 and idaction = 2 and these
actions have transition probabilities associated with them, and idaction = 1 has
no transition probabilities, then idaction = 2 is renamed to idaction = 1.
}
\details{
The result contains a mapping (action_mapping) dataframe that for
each state (idstate) contains an idaction_old which is the action before packing
and an idaction_new which is the id of the action after packing
}
