% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcraam_set_threads}
\alias{rcraam_set_threads}
\title{Sets the number of threads for parallelization.}
\usage{
rcraam_set_threads(n)
}
\arguments{
\item{n}{Number of threads}
}
\description{
Sets the number of threads for parallelization.
}
