% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcraam_supports_gurobi}
\alias{rcraam_supports_gurobi}
\title{Whether Gurobi LP and MILP is installed}
\usage{
rcraam_supports_gurobi()
}
\description{
This function can be used when determining which functionality
is available in the package
}
