% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{revaluate_mdpo_rnd}
\alias{revaluate_mdpo_rnd}
\title{Evaluates a randomized policy for many Bayesian samples}
\usage{
revaluate_mdpo_rnd(mdpo, discount, policy_rnd, initial, show_progress = TRUE)
}
\arguments{
\item{mdpo}{Dataframe with `idstatefrom`, `idaction`, `idstateto`, `idoutcome`, `probability`, `reward`.
Each `idoutcome` represents a sample. The outcomes must be sorted increasingly.}

\item{discount}{Discount rate in [0,1) (or = 1 at the risk of divergence)}

\item{initial}{Initial distribution with columns `idstate` and `probability`}

\item{show_progress}{Whether to show a progress bar}

\item{policy_rand}{Randomized policy with columns `idstate`, `idaction`, `probability`}
}
\value{
List of returns for all outcomes
}
\description{
Evaluates a randomized policy for many Bayesian samples
}
