% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srsolve_mdpo}
\alias{srsolve_mdpo}
\title{Solves an MDPO with static uncertainty using a non-convex global optimization method.}
\usage{
srsolve_mdpo(
  mdpo,
  init_distribution,
  discount,
  alpha,
  beta,
  algorithm = "quadratic",
  model_distribution = NULL,
  output_filename = ""
)
}
\arguments{
\item{mdpo}{Uncertain MDP. The outcomes are assumed to represent the uncertainty over MDPs.
The number of outcomes must be uniform for all states and actions
(except for terminal states which have no actions).}

\item{init_distribution}{Initial distribution over states. The columns should be
are idstate, and probability.}

\item{alpha}{Risk level of avar (0 = worst-case). The minimum value is 1e-5, the maximum
value is 1.}

\item{beta}{Weight on AVaR and the complement (1-beta) is the weight
on the expectation term. The value must be between 0 and 1.}

\item{model_distribution}{Distribution over the models. The default is empty, which translates
to a uniform distribution. The columns should be idstate, and probablity.}

\item{output_filename}{Name of the file to save the model output. Valid suffixes are
  .mps, .rew, .lp, or .rlp for writing the model itself.
If it is an empty string, then it does not write the file.}

\item{gamma}{Discount factor. Clamped to be in [0,1]}
}
\description{
The objective is:
 \deqn{\max_{\pi} \beta * CVaR_{P \sim f}^\alpha [return(\pi,P)] + (1-\beta) * E_{P \sim f}^\alpha [return(\pi,P)]}
}
