% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{worstcase_l1_w}
\alias{worstcase_l1_w}
\title{Computes the maximum distribution subject to weighted L1 constraints}
\usage{
worstcase_l1_w(value, reference_dst, w, budget)
}
\arguments{
\item{value}{Random variable (objective)}

\item{reference_dst}{Reference distribution of the same size as value}

\item{w}{set of weights for ambiguity set}

\item{budget}{Maximum L1 distance from the reference dst}
}
\value{
A list with dst as the worstcase distribution,
        and value as the objective
}
\description{
Computes the maximum distribution subject to weighted L1 constraints
}
