% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{worstcase_linf_w_gurobi}
\alias{worstcase_linf_w_gurobi}
\title{Computes the maximum distribution subject to weighted Linf constraints using Gurobi}
\usage{
worstcase_linf_w_gurobi(value, reference_dst, w, budget)
}
\arguments{
\item{value}{Random variable (objective)}

\item{reference_dst}{Reference distribution of the same size as value}

\item{w}{set of weights for ambiguity set}

\item{budget}{Maximum Linf distance from the reference dst}
}
\value{
A list with dst as the worstcase distribution,
        and value as the objective
}
\description{
The function is only supported when the package is installed with Gurobi support
}
